Reimagining the whole thing to make it simpler, plus writing an UciCodec to map between byte arrays and UciMessages.
