+++ Cargo.toml
@@ -17,9 +17,10 @@ travis-ci = { repository = "vampirc/vampirc-io", branch = "master" }
 maintenance = { status = "actively-developed" }
 
 [features]
-default = ["vampirc-uci"]
+#default = ["vampirc-uci"]
 
 [dependencies]
 tokio = "0.1"
+tokio-codec = "0.1"
 bytes = "0.4"
-vampirc-uci = { git = 'https://github.com/vampirc/vampirc-uci.git', optional=true}
+vampirc-uci = { git = 'https://github.com/vampirc/vampirc-uci.git'}
